+++ src/settings.spec.ts
@@ -1,17 +1,8 @@
-/**
+/*
  * Copyright (c) 2018-present, Matei Bogdan Radu <matei.radu.92@gmail.com>
  *
  * This source code is licensed under the MIT license found in the LICENSE
  * file in the root directory of this source tree.
- */
-
-/**
- * Since ts-jest v23.10.x, full type-checking is performed in TS test files, so
- * tests would fail for incompatible assignments.
- *
- * This library can be used in plain Javascript project, so behaviour when
- * incompatible assignments are made must be evaluated. To do so, this file
- * must be kept in Javascript to correctly simulate that scenario.
  */
 
 import { sanitize, initialize, defaultSettings } from "./settings";
@@ -48,7 +39,7 @@ describe("Sanitize settings - Android", () => {
     expect(
       sanitize("android", {
         android: { addDefaultShareMenu: null }
-      })
+      } as any)
     ).toEqual({});
   });
 
@@ -67,9 +58,9 @@ describe("Sanitize settings - Android", () => {
           toolbarColor: "#3fF",
           showTitle: null
         }
-      })
+      } as any)
     ).toEqual({
-      toolbarColor: "#33ffFF"
+      toolbarColor: "#33ffff"
     });
   });
 
@@ -88,7 +79,7 @@ describe("Sanitize settings - Android", () => {
         }
       })
     ).toEqual({
-      toolbarColor: "red",
+      toolbarColor: "#ff0000",
       showTitle: true
     });
   });
@@ -122,7 +113,7 @@ describe("Sanitize settings - iOS", () => {
     expect(
       sanitize("ios", {
         ios: { barCollapsingEnabled: null }
-      })
+      } as any)
     ).toEqual({});
   });
 
@@ -140,7 +131,7 @@ describe("Sanitize settings - iOS", () => {
         }
       })
     ).toEqual({
-      preferredBarTintColor: "#33ffFF",
+      preferredBarTintColor: "#33ffff",
       barCollapsingEnabled: true
     });
   });
@@ -161,8 +152,8 @@ describe("Sanitize settings - iOS", () => {
         }
       })
     ).toEqual({
-      preferredBarTintColor: "#33ffFF",
-      preferredControlTintColor: "#33ffFF",
+      preferredBarTintColor: "#33ffff",
+      preferredControlTintColor: "#33ffff",
       barCollapsingEnabled: true
     });
   });
@@ -189,10 +180,10 @@ describe("Initialize", () => {
 
     expect(defaultSettings).toEqual({
       android: {
-        toolbarColor: "red"
+        toolbarColor: "#ff0000"
       },
       ios: {
-        preferredBarTintColor: "#33ffFF"
+        preferredBarTintColor: "#33ffff"
       }
     });
   });