+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.7] - 2019-01-29
+## [1.3.0] - 2019-02-27
+
+### Added
+
+- Example project in the `example` directory.
+
+### Changed
+
+- Color properties (`toolbarColor`, `preferredBarTintColor` and `preferredControlTintColor`) now
+  accept more color formats. Please refer to TinyColor's [Accepted String Input](https://github.com/bgrins/TinyColor#accepted-string-input) to see all available formats.
+- The iOS `.podspec` will now automatically stay in sync with `package.json` (thanks to [@ericlewis](https://github.com/ericlewis)).
+
+## [1.2.7] - 2019-02-21
 
 ### Added
 
@@ -85,6 +97,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `openInApp`: open a valid http(s) URL with an in-app browser.
 
+[1.3.0]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.2.7...v1.3.0
 [1.2.7]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.2.4...v1.2.7
 [1.2.4]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.2.1...v1.2.4
 [1.2.1]: https://github.com/matt-block/react-native-in-app-browser/compare/v1.2.0...v1.2.1